Update `BloodSugarMeasurements` DB schema
- Change `BloodSugarReading#value` type from `Int` to `Float`

+++ app/src/main/java/org/simple/clinic/bloodsugar/sync/BloodSugarMeasurementPayload.kt
@@ -46,7 +46,7 @@ data class BloodSugarMeasurementPayload(
   fun toDatabaseModel(syncStatus: SyncStatus) = BloodSugarMeasurement(
       uuid = uuid,
       reading = BloodSugarReading(
-          value = bloodSugarValue,
+          value = bloodSugarValue.toFloat(),
           type = bloodSugarType
       ),
       syncStatus = syncStatus,
